changes per review comments
