+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/events/QueuedEvent.java
@@ -19,7 +19,7 @@ public class QueuedEvent {
   private final EventType eventType;
   private final String messageData;
 
-  public QueuedEvent(final EventType eventType, final String messageData) {
+  private QueuedEvent(final EventType eventType, final String messageData) {
     this.eventType = eventType;
     this.messageData = messageData;
   }
